requirements.txt
# local package
-e .

# external requirements
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
croniter==1.3.15
cycler==0.11.0
databricks-cli==0.17.7
dateutils==0.6.12
deepdiff==6.3.0
docker==6.1.2
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.88.0
filelock==3.9.0
Flask==2.3.2
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
kiwisolver==1.4.4
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mlflow==2.3.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
Pillow==9.3.0
protobuf==4.23.2
psutil==5.9.5
pyarrow==11.0.0
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
pywin32==306
PyYAML==6.0
querystring-parser==1.2.4
readchar==4.0.5
requests==2.28.1
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==0.11.4
torchvision==0.15.2+cu118
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.22.0
waitress==2.1.2
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
